refactor(cart): streamline cart service methods and enhance validation logic
